requirements.txt
absl-py==0.1.11
astor==0.6.2
bleach==1.5.0
certifi==2018.1.18
colorama==0.3.9
cycler==0.10.0
decorator==4.2.1
entrypoints==0.2.3
gast==0.2.0
grpcio==1.10.0
h5py==2.7.1
html5lib==0.9999999
ipykernel==4.8.2
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.2
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
kaggle==1.0.5
Keras==2.1.5
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.0
numpy==1.14.1
opencv-python==3.4.0.12
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
protobuf==3.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2018.3
pywinpty==0.5.1
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.3.1
scipy==1.0.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
tensorboard==1.6.0
tensorflow-gpu==1.6.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
tornado==5.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.14.1
widgetsnbextension==3.1.4
wincertstore==0.2
